Refactor helpers and warning
