+++ src/meteofrance/model/__init__.py
@@ -2,3 +2,4 @@
 from .forecast import Forecast
 from .place import Place
 from .rain import Rain
+from .warning import CurrentPhenomenons, Full
