+++ tests/test_warning.py
@@ -4,12 +4,6 @@ import pytest
 
 from meteofrance.auth import MeteoFranceAuth
 from meteofrance.client import MeteoFranceClient
-from meteofrance.warning import (
-    get_phenomenon_name_from_indice,
-    get_text_status_from_indice_color,
-    is_coastal_department,
-    readeable_phenomenoms_dict,
-)
 
 WARNING_COLOR_LIST = [1, 2, 3, 4]
 
@@ -62,44 +56,6 @@ def test_thumbnail():
         "?&token=__Wj7dVSTjV9YGu1guveLyDq0g7S7TfTjaHBTPTpO0kj8__&"
         "domain=france"
     )
-
-
-def test_text_helpers_fr():
-    """Test helpers to have readable alert type and alert level in French."""
-    assert get_text_status_from_indice_color(1) == "Vert"
-    assert get_phenomenon_name_from_indice(2) == "Pluie-inondation"
-
-
-def test_get_text_status_from_indice_color_en():
-    """Test helpers to have readable alert type and alert level in English."""
-    assert get_text_status_from_indice_color(4, "en") == "Red"
-    assert get_phenomenon_name_from_indice(4, "en") == "Flood"
-
-
-@pytest.mark.parametrize("dep, res", [("03", False), ("06", True), ("2B", True)])
-def test_is_coastal_department(dep, res):
-    """Test the helper checking if an additional coastal departement bulletin exist."""
-    assert is_coastal_department(dep) == res
-
-
-def test_readeable_phenomenoms_dict():
-    """Test the helper constructing a human readable dictionary for phenomenom."""
-    api_list = [
-        {"phenomenon_id": 4, "phenomenon_max_color_id": 1},
-        {"phenomenon_id": 5, "phenomenon_max_color_id": 1},
-        {"phenomenon_id": 3, "phenomenon_max_color_id": 2},
-        {"phenomenon_id": 2, "phenomenon_max_color_id": 1},
-        {"phenomenon_id": 1, "phenomenon_max_color_id": 3},
-    ]
-    expected_dictionary = {
-        "Inondation": "Vert",
-        "Neige-verglas": "Vert",
-        "Pluie-inondation": "Vert",
-        "Orages": "Jaune",
-        "Vent violent": "Orange",
-    }
-
-    assert readeable_phenomenoms_dict(api_list) == expected_dictionary
 
 
 @pytest.mark.parametrize("dep, res", [("13", True), ("32", False)])
